`Rename "wishlist.strore" to "wishlist.store" in multiple files`
